stripe code added

+++ database/migrations/2026_01_01_062609_add_slug_to_products_table.php
@@ -17,7 +17,9 @@ return new class extends Migration {
     public function down(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('slug');
+            if (Schema::hasColumn('products', 'slug')) {
+                $table->dropColumn('slug');
+            }
         });
     }
 };
